+++ src/main/java/net/rpgz/mixin/HopperBlockEntityMixin.java
@@ -1,9 +1,10 @@
 package net.rpgz.mixin;
 
-import java.util.Iterator;
 import java.util.List;
 
+import net.minecraft.entity.mob.MobEntity;
 import net.minecraft.world.World;
+import net.rpgz.util.RpgHelper;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -12,17 +13,15 @@ import org.spongepowered.asm.mixin.injection.At;
 
 import net.minecraft.block.entity.Hopper;
 import net.minecraft.block.entity.HopperBlockEntity;
-import net.minecraft.entity.LivingEntity;
 import net.minecraft.inventory.Inventory;
 import net.minecraft.predicate.entity.EntityPredicates;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Box;
 import net.minecraft.util.math.Direction;
-import net.rpgz.access.InventoryAccess;
 import net.rpgz.init.ConfigInit;
 
 @Mixin(HopperBlockEntity.class)
-public abstract class HopperBlockEntityMixin implements InventoryAccess {
+public abstract class HopperBlockEntityMixin {
     private static int ticking = 0;
 
     @Inject(method = "extract(Lnet/minecraft/world/World;Lnet/minecraft/block/entity/Hopper;)Z", at = @At(value = "INVOKE", target = "Lnet/minecraft/block/entity/HopperBlockEntity;getInputItemEntities(Lnet/minecraft/world/World;Lnet/minecraft/block/entity/Hopper;)Ljava/util/List;"), cancellable = true)
@@ -32,21 +31,16 @@ public abstract class HopperBlockEntityMixin implements InventoryAccess {
             if (ticking >= 20) {
                 BlockPos pos = BlockPos.ofFloored(hopper.getHopperX(), hopper.getHopperY(), hopper.getHopperZ());
                 Box box = new Box(pos).expand(0.0D, 1.0D, 0.0D);
-                List<LivingEntity> list = world.getEntitiesByClass(LivingEntity.class, box, EntityPredicates.EXCEPT_SPECTATOR);
+                List<MobEntity> list = world.getEntitiesByClass(MobEntity.class, box, EntityPredicates.EXCEPT_SPECTATOR);
                 if (!list.isEmpty()) {
-                    Iterator<LivingEntity> iterator = list.iterator();
-                    while (iterator.hasNext()) {
-                        LivingEntity livingEntity = (LivingEntity) iterator.next();
-                        if (livingEntity.isDead()) {
-                            if (((InventoryAccess) livingEntity).getInventory() != null) {
-                                Direction direction = Direction.DOWN;
-
-                                for (int i : getAvailableSlots(((InventoryAccess) livingEntity).getInventory(), direction)) {
-                                    if (!extract(hopper, ((InventoryAccess) livingEntity).getInventory(), i, direction)) {
-                                        continue;
-                                    }
-                                    info.setReturnValue(true);
+                    for (MobEntity mobEntity : list) {
+                        if (mobEntity.isDead()) {
+                            Direction direction = Direction.DOWN;
+                            for (int i : getAvailableSlots(RpgHelper.getDeadMobInventory(mobEntity), direction)) {
+                                if (!extract(hopper, RpgHelper.getDeadMobInventory(mobEntity), i, direction)) {
+                                    continue;
                                 }
+                                info.setReturnValue(true);
                             }
                         }
                     }